+++ src/components/index.js
@@ -230,7 +230,7 @@ export {
 
 export {
   EuiNavDrawer,
-  EuiNavDrawerMenu,
+  EuiNavDrawerGroup,
   EuiNavDrawerFlyout,
 } from './nav_drawer';