+++ src/components/list_group/list_group.js
@@ -15,6 +15,7 @@ export const EuiListGroup = ({
   listItems,
   maxWidth,
   style,
+  showToolTips,
   ...rest,
 }) => {
 
@@ -32,7 +33,6 @@ export const EuiListGroup = ({
     {
       'euiListGroup-flush': flush,
       'euiListGroup-bordered': bordered,
-      'euiListGroup-wrapText': wrapText,
     },
     widthClassName,
     className
@@ -45,13 +45,23 @@ export const EuiListGroup = ({
         return [
           <EuiListGroupItem
             key={`title-${index}`}
+            showToolTip={showToolTips}
+            wrapText={wrapText}
             {...item}
           />
         ];
       })
     );
   } else {
-    childrenOrListItems = children;
+    if (showToolTips) {
+      childrenOrListItems = React.Children.map(children, child => {
+        return React.cloneElement(child, {
+          showToolTip: true
+        });
+      });
+    } else {
+      childrenOrListItems = children;
+    }
   }
 
   return (
@@ -73,6 +83,7 @@ EuiListGroup.propTypes = {
     iconType: PropTypes.string,
     isActive: PropTypes.boolean,
     isDisabled: PropTypes.boolean,
+    showToolTip: PropTypes.boolean,
   })),
   children: PropTypes.node,
   className: PropTypes.string,
@@ -93,6 +104,11 @@ EuiListGroup.propTypes = {
   wrapText: PropTypes.bool,
 
   /**
+   * Display tooltips on all list items
+   */
+  showToolTips: PropTypes.bool,
+
+  /**
    * Sets the max-width of the page,
    * set to `true` to use the default size,
    * set to `false` to not restrict the width,
@@ -111,4 +127,5 @@ EuiListGroup.defaultProps = {
   bordered: false,
   wrapText: false,
   maxWidth: true,
+  showToolTips: false,
 };